pyproject.toml
[project]
name = "funeralthebot"
version = "0.1.0"
description = "Bot for Telegram"
authors = [
    {name = "dimitrioglo.legacy@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"


[tool.poetry]
package-mode = false


[tool.poetry.dependencies]
python-telegram-bot = "^22.4"
python-dotenv = "^1.1.1"
telethon = "^1.41.2"
async-lru = "^2.0.5"
dependency-injector = "^4.48.2"
uvicorn = "^0.37.0"
starlette = "^0.48.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.13.1"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 80

[tool.ruff.lint]
select = [
    "I",
    "E4",
    "E7",
    "E9",
    "F",
    "B",
    "UP"
]
